Round forecast SoC to whole percent

+++ custom_components/energy_planner/models.py
@@ -30,7 +30,7 @@ class SolarForecastPoint:
 @dataclass(frozen=True)
 class SocForecastPoint:
     timestamp: datetime
-    soc_percent: float
+    soc_percent: int
     battery_kwh: float
     solar_kwh: float
     consumption_kwh: float
